feat(admin): add settings and add-photo page to expand admin functions
