+++ app/admin/page.module.scss
@@ -1,48 +1,52 @@
 @use '@/styles/vars' as *;
 
-.AddPhoto {
-  .content {
-    max-width: 80%;
-    width: 600px;
-    margin: $gap-2xl auto;
-
-    input,
-    textarea,
-    select {
-      padding: $gap-regular;
-      border-radius: 8px;
-      outline: none;
-      border: 1px solid $border-color;
+.adminPage {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+  width: 100%;
+  .links {
+    margin-top: $gap-3xl;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    gap: $gap-medium;
+    width: fit-content;
+    a {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      gap: $gap-small;
     }
+  }
+}
+
+.adminNav {
+  position: fixed;
+  top: 24px;
+  left: 24px;
+}
+
+.settingsPage {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+  width: 100%;
+  margin-top: $gap-3xl;
+  .buttons {
+    display: flex;
+    flex-direction: column;
+    gap: $gap-large;
     button {
       border: 1px solid $border-color;
-      margin: 0 auto;
-      margin-top: $gap-regular;
-      width: fit-content;
-      padding: 16px 24px;
+      padding: $gap-small $gap-medium;
       border-radius: 16px;
-    }
-
-    .fileInputLabel {
-      display: block;
-      width: 100%;
-      padding: 1rem;
-      border: 2px dashed #aaa;
-      border-radius: 8px;
-      text-align: center;
-      color: #666;
-      font-weight: 600;
-      cursor: pointer;
-      transition: border-color 0.3s ease, color 0.3s ease;
-      user-select: none;
-
       &:hover {
-        border-color: #333;
-        color: #000;
-      }
-
-      input[type='file'] {
-        display: none;
+        background: red;
+        color: white;
       }
     }
   }
